.old/back/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
geojson==2.5.0
idna==2.10
iniconfig==1.1.1
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
requests==2.25.1
rply==0.7.8
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
